+++ example_test.go
@@ -9,18 +9,18 @@ func Example() {
 		Address: url,
 	})
 	if err != nil {
-		log.Fatalln(err.Error())
+		Log.Fatalln(err.Error())
 	}
 
 	res, err := Expr("Hello World").Run(session)
 	if err != nil {
-		log.Fatalln(err.Error())
+		Log.Fatalln(err.Error())
 	}
 
 	var response string
 	err = res.One(&response)
 	if err != nil {
-		log.Fatalln(err.Error())
+		Log.Fatalln(err.Error())
 	}
 
 	fmt.Println(response)