+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
  - Added `Union` as a root term
  - Added `Branch` as a root term
  - Added `ReadTimeout` and `WriteTimeout` to `RunOpts` and `ExecOpts`
+ - Exported `github.com/Sirupsen/logrus.Logger` as `Log`
 
 ### Changed
  - Changed `Timeout` connect argument to only configure the connection timeout.